Fixing Capabilities Message

+++ modules/uas_message/capabilities_message.cpp
@@ -31,9 +31,9 @@ UASMessage::MessageID CapabilitiesMessage::type()
 std::vector<unsigned char> CapabilitiesMessage::serialize()
 {
     std::vector<unsigned char> serializedMessage;
-    serializedMessage.push_back(static_cast<uint8_t>(capabilities >> 24) & 0xFF);
-    serializedMessage.push_back(static_cast<uint8_t>(capabilities >> 16) & 0xFF);
-    serializedMessage.push_back(static_cast<uint8_t>(capabilities >> 8) & 0xFF);
-    serializedMessage.push_back(static_cast<uint8_t>(capabilities) & 0xFF);
+    serializedMessage.push_back((static_cast<uint32_t>(capabilities) >> 24) & 0xFF);
+    serializedMessage.push_back((static_cast<uint32_t>(capabilities) >> 16) & 0xFF);
+    serializedMessage.push_back((static_cast<uint32_t>(capabilities) >> 8) & 0xFF);
+    serializedMessage.push_back((static_cast<uint32_t>(capabilities)) & 0xFF);
     return serializedMessage;
 }
